+++ CHANGELOG.md
@@ -5,12 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [NEXT] 
-- Added an Box for Information on the Homescreen [#216] (https://github.com/h-da/geli/issues/216)
-
 ## [0.6.0] - 2018-03-31 - Introduces MediaManager and some minor changes
 ### Added
 - MediaManager for file management in courses
-
+- Added an Box for Information on the Homescreen [#216] (https://github.com/h-da/geli/issues/216)
 
 ## [0.5.0] - 2018-03-24 - WS 17/18 intermediate Release
 ### Added